fix(plugins): keep the tooltip graph off every panel's first render
CI caught two things targeted local runs could not.

- `plugin.handlers.test.ts` asserts an exact handler count, which this branch
  moved from 53 to 55 with `plugin:runtime-statuses-get` and
  `plugin:restart-worker`. Both are now named in the test so the count says
  what it is counting.
- The runtime-status layer statically imported `InlineStatusBanner`, which
  pulls the tooltip and window-chrome graph behind it. That layer renders
  inside EVERY plugin panel, so the graph landed on the first-render path of
  every healthy panel that will never show a banner — and in a suite that
  documents mounting the content layer without any of that chrome. The banner
  is now its own lazily-loaded module, and the presentation logic
  (`presentWorkerStatus`, `useWorkerStall`) sits in a dependency-free module
  the content layer imports directly.
